added lambda example for Neptune
