+++ javav2/example_code/neptune/src/main/java/com/example/neptune/database/NeptuneGremlinQueryExample.java
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.example.neptune.data;
+package com.example.neptune.database;
 
 import software.amazon.awssdk.core.client.config.ClientOverrideConfiguration;
 import software.amazon.awssdk.http.apache.ApacheHttpClient;
